--- a/ledger-service/http-json-perf/BUILD.bazel
+++ b/ledger-service/http-json-perf/BUILD.bazel
@@ -32,8 +32,8 @@ da_scala_library(
         "//ledger-service/http-json-testing",
         "//ledger-service/jwt",
         "//libs-scala/gatling-utils",
-        "//libs-scala/scala-utils",
         "//libs-scala/postgresql-testing",
+        "//libs-scala/scala-utils",
         "@maven//:com_fasterxml_jackson_core_jackson_core",
         "@maven//:com_fasterxml_jackson_core_jackson_databind",
         "@maven//:com_github_scopt_scopt_2_12",
@@ -46,10 +46,10 @@ da_scala_library(
         "@maven//:io_gatling_gatling_core",
         "@maven//:io_gatling_gatling_http",
         "@maven//:io_gatling_gatling_http_client",
-        "@maven//:org_scalaz_scalaz_core_2_12",
-        "@maven//:org_slf4j_slf4j_api",
         "@maven//:org_scalactic_scalactic_2_12",
         "@maven//:org_scalatest_scalatest_2_12",
+        "@maven//:org_scalaz_scalaz_core_2_12",
+        "@maven//:org_slf4j_slf4j_api",
     ],
 )
 
@@ -92,10 +92,10 @@ da_scala_binary(
         "@maven//:io_gatling_gatling_core",
         "@maven//:io_gatling_gatling_http",
         "@maven//:io_gatling_gatling_http_client",
-        "@maven//:org_scalaz_scalaz_core_2_12",
-        "@maven//:org_slf4j_slf4j_api",
         "@maven//:org_scalactic_scalactic_2_12",
         "@maven//:org_scalatest_scalatest_2_12",
+        "@maven//:org_scalaz_scalaz_core_2_12",
+        "@maven//:org_slf4j_slf4j_api",
     ],
 )