--- a/ledger-service/http-json-perf/src/main/scala/com/daml/http/perf/Config.scala
+++ b/ledger-service/http-json-perf/src/main/scala/com/daml/http/perf/Config.scala
@@ -18,15 +18,17 @@ private[perf] final case class Config[+S](
     dars: List[File],
     jwt: Jwt,
     reportsDir: File,
-    maxDuration: Option[FiniteDuration]
+    maxDuration: Option[FiniteDuration],
+    sandboxPersistence: Boolean
 ) {
   override def toString: String =
     s"Config(" +
-      s"scenario=${this.scenario}, " +
-      s"dars=${dars: List[File]}," +
-      s"jwt=..., " + // don't print the JWT
-      s"reportsDir=${reportsDir: File}," +
-      s"maxDuration=${this.maxDuration: Option[FiniteDuration]}" +
+      s"scenario=${this.scenario}" +
+      s", dars=${dars: List[File]}, " +
+      s", jwt=..." + // don't print the JWT
+      s", reportsDir=${reportsDir: File}" +
+      s", maxDuration=${this.maxDuration: Option[FiniteDuration]}" +
+      s", sandboxPersistence=${this.sandboxPersistence: Boolean}" +
       ")"
 }
 
@@ -37,7 +39,8 @@ private[perf] object Config {
       dars = List.empty,
       jwt = Jwt(""),
       reportsDir = new File(""),
-      maxDuration = None)
+      maxDuration = None,
+      sandboxPersistence = false)
 
   implicit val configInstance: Traverse[Config] = new Traverse[Config] {
     override def traverseImpl[G[_]: Applicative, A, B](fa: Config[A])(
@@ -74,6 +77,11 @@ private[perf] object Config {
         .required()
         .validate(validateJwt)
         .text("JWT token to use when connecting to JSON API.")
+
+      opt[Boolean]("sandbox-persistence")
+        .action((x, c) => c.copy(sandboxPersistence = x))
+        .required()
+        .text("Enable or disable sandbox persistence.")
 
       opt[File]("reports-dir")
         .action((x, c) => c.copy(reportsDir = x))