--- a/ledger-service/http-json-perf/src/main/scala/com/daml/http/perf/PostgresRunner.scala
+++ b/ledger-service/http-json-perf/src/main/scala/com/daml/http/perf/PostgresRunner.scala
@@ -1,0 +1,13 @@
+package com.daml.http.perf
+
+import com.daml.testing.postgresql.PostgresAroundSuite
+import org.scalatest.Suite
+
+class PostgresRunner {
+  private val suite = new PostgresAroundSuite with Suite {}
+
+  def createNewRandomDatabase(): Unit = {
+//    suite.c
+  }
+
+}