--- a/ledger-service/http-json-perf/src/main/scala/com/daml/http/perf/PostgresRunner.scala
+++ b/ledger-service/http-json-perf/src/main/scala/com/daml/http/perf/PostgresRunner.scala
@@ -1,13 +1,28 @@
+// Copyright (c) 2020 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
 package com.daml.http.perf
 
-import com.daml.testing.postgresql.PostgresAroundSuite
+import com.daml.testing.postgresql.{PostgresAroundSuite, PostgresDatabase}
 import org.scalatest.Suite
 
+import scala.util.Try
+
 class PostgresRunner {
-  private val suite = new PostgresAroundSuite with Suite {}
 
-  def createNewRandomDatabase(): Unit = {
-//    suite.c
-  }
+  private val suite = new PostgresRunnerSuite()
 
+  def start(): Try[PostgresDatabase] = suite.start()
+
+  def stop(): Try[Unit] = suite.stop()
+}
+
+private class PostgresRunnerSuite extends PostgresAroundSuite with Suite {
+  def start(): Try[PostgresDatabase] =
+    for {
+      _ <- Try(this.connectToPostgresqlServer())
+      db <- Try(this.createNewDatabase())
+    } yield db
+
+  def stop(): Try[Unit] = Try(this.disconnectFromPostgresqlServer())
 }