Closes: #7167

### Notes
Query Store affects only sync query scenarios hitting `/v1/query` endpoint:
- `com.daml.http.perf.scenario.SyncQueryConstantAcs`
- `com.daml.http.perf.scenario.SyncQueryNewAcs`
- `com.daml.http.perf.scenario.SyncQueryVariableAcs`

### Running with query store index (--query-store-index=true):
```
$ bazel run //ledger-service/http-json-perf:http-json-perf-binary -- \
 --query-store-index=true \
 --scenario=com.daml.http.perf.scenario.SyncQueryConstantAcs \
 --dars=$(pwd)/bazel-bin/docs/quickstart-model.dar \
 --reports-dir=/var/tmp/results \
 --jwt="eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJodHRwczovL2RhbWwuY29tL2xlZGdlci1hcGkiOnsibGVkZ2VySWQiOiJNeUxlZGdlciIsImFwcGxpY2F0aW9uSWQiOiJmb29iYXIiLCJhY3RBcyI6WyJBbGljZSJdfX0.VdDI96mw5hrfM5ZNxLyetSVwcD7XtLT4dIdHIOa9lcU"
```
```
================================================================================
---- SyncQueryRequest ----------------------------------------------------------
> Number of requests                                   500 (OK=500    KO=-     )
> Min. response time                                    17 (OK=17     KO=-     )
> Max. response time                                  1546 (OK=1546   KO=-     )
> Mean response time                                    27 (OK=27     KO=-     )
> Std. deviation                                        68 (OK=68     KO=-     )
> response time 90th percentile                         31 (OK=31     KO=-     )
> response time 95th percentile                         37 (OK=37     KO=-     )
> response time 99th percentile                         51 (OK=51     KO=-     )
> response time 99.9th percentile                     1546 (OK=1546   KO=-     )
> Mean requests/second                              36.724 (OK=36.724 KO=-     )
---- Response time distribution ------------------------------------------------
> t < 5000 ms                                          500 (100%)
> 5000 ms < t < 30000 ms                                 0 (  -%)
> 30000 ms < t                                           0 (  -%)
> failed                                                 0 (  -%)
================================================================================
```
### Running without query store index (--query-store-index=false):
```
$ bazel run //ledger-service/http-json-perf:http-json-perf-binary -- \
 --query-store-index=false \
 --scenario=com.daml.http.perf.scenario.SyncQueryConstantAcs \
 --dars=$(pwd)/bazel-bin/docs/quickstart-model.dar \
 --reports-dir=/var/tmp/results \
 --jwt="eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJodHRwczovL2RhbWwuY29tL2xlZGdlci1hcGkiOnsibGVkZ2VySWQiOiJNeUxlZGdlciIsImFwcGxpY2F0aW9uSWQiOiJmb29iYXIiLCJhY3RBcyI6WyJBbGljZSJdfX0.VdDI96mw5hrfM5ZNxLyetSVwcD7XtLT4dIdHIOa9lcU"
```
```
================================================================================
---- SyncQueryRequest ----------------------------------------------------------
> Number of requests                                   500 (OK=500    KO=-     )
> Min. response time                                   175 (OK=175    KO=-     )
> Max. response time                                   874 (OK=874    KO=-     )
> Mean response time                                   208 (OK=208    KO=-     )
> Std. deviation                                        35 (OK=35     KO=-     )
> response time 90th percentile                        222 (OK=222    KO=-     )
> response time 95th percentile                        234 (OK=234    KO=-     )
> response time 99th percentile                        288 (OK=288    KO=-     )
> response time 99.9th percentile                      874 (OK=874    KO=-     )
> Mean requests/second                               4.809 (OK=4.809  KO=-     )
---- Response time distribution ------------------------------------------------
> t < 5000 ms                                          500 (100%)
> 5000 ms < t < 30000 ms                                 0 (  -%)
> 30000 ms < t                                           0 (  -%)
> failed                                                 0 (  -%)
================================================================================
```

### Pull Request Checklist

- [ ] Read and understand the [contribution guidelines](https://github.com/digital-asset/daml/blob/master/CONTRIBUTING.md)
- [ ] Include appropriate tests
- [ ] Set a descriptive title and thorough description
- [ ] Add a reference to the [issue this PR will solve](https://github.com/digital-asset/daml/issues), if appropriate
- [ ] Include changelog additions in one or more commit message bodies between the `CHANGELOG_BEGIN` and `CHANGELOG_END` tags
- [ ] Normal production system change, include purpose of change in description

NOTE: CI is not automatically run on non-members pull-requests for security
reasons. The reviewer will have to comment with `/AzurePipelines run` to
trigger the build.